Refactor log logic
Added functionality to calculate each log's 1 rep max which is then stored when the the log is created. Reworked log fetching so that logs are displayed only for that exercise with the most recent on top. Added functionality for the user to click on an existing log to populate the form for quicker entry.